Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2: -Code-Review
> > > 
> > > yeah topic isnt needed, no one should be doing this in jenkins
> > 
> > Hey, i wanted to become official and i just applied. I'm already doing unoffocial builds and i use qcom fbe wrapped key encryption. That means that i should drop it if i want to make official or community builds (if i'm accepted of course).
> > 
> > What's wrong with this encryption and why forbid to use it ?
> 
> current wrappedkey patches port are WIP and do not work.

Ok thx for your answer. Then i will have to drop it to become official and users would have to clean flash. And once it's properly working i will change again for it.

The thing is that i currently use it and i don't see any issue

Patch-set: 2